track_devices/server/requirements.txt
astroid==2.3.3
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
grpcio==1.27.2
grpcio-tools==1.27.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
protobuf==3.6.1
psutil==5.7.0
py-ubjson==0.14.0
pylint==2.4.4
PyMySQL==0.9.3
PyYAML==5.3
simplejson==3.17.0
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
